数仓表结构不符合报错信息更改 模式引入删除不存在属性增加判断 提示sql查询报错信息

+++ kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/GraphMapServiceImpl.java
@@ -342,11 +342,15 @@ public class GraphMapServiceImpl implements GraphMapService {
                     //图谱中该属性的边属性不存在，删除订阅记录
                     for(DWGraphMapRelationAttr relationAttr : relationAttrList){
                         if(!relationAttrNames.contains(relationAttr.getName())){
-                            graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            if(graphMapRelationAttrRepository.existsById(relationAttr.getId())){
+                                graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            }
                         }
 
                         if(existRelationAttList.contains(relationAttr.getName()+relationAttr.getModelId()+relationAttr.getModelAttrId())){
-                            graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            if(graphMapRelationAttrRepository.existsById(relationAttr.getId())) {
+                                graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            }
                         }else{
                             existRelationAttList.add(relationAttr.getName()+relationAttr.getModelId()+relationAttr.getModelAttrId());
                         }
